Move Output to analyzer package
