+++ presto-main/src/test/java/io/prestosql/sql/analyzer/TestOutput.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.execution;
+package io.prestosql.sql.analyzer;
 
 import io.airlift.json.JsonCodec;
 import io.prestosql.connector.CatalogName;
